Moving theme into ui package.
