+++ app/src/main/java/com/adammcneilly/pokedex/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import com.adammcneilly.pokedex.theme.DexTheme
+import com.adammcneilly.pokedex.ui.theme.DexTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
